+++ database/migrations/business/registry.go
@@ -15,6 +15,7 @@ type Migration struct {
 	ID                string
 	Revision          uint64
 	Up                func(*gorm.DB, *conf.Configuration) error
+	Down              func(*gorm.DB, *conf.Configuration) error
 	VerifyBaseline    func(*gorm.DB, *conf.Configuration) error
 	VerifySchema      func(*gorm.DB, *conf.Configuration) error
 	VerifyUpgradeData func(*gorm.DB, *conf.Configuration) error